backend/pyproject.toml
[project]
name = "devlog-backend"
version = "1.0.0"
description = "DevLog AI Backend - Embedding, AI, and Agent Services"
requires-python = ">=3.10,<3.14"
dependencies = [
    # FastAPI and server
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    
    # Embedding model
    "sentence-transformers>=3.0.0",
    "torch>=2.1.0",
    "numpy>=1.24.0,<2.3.0",
    "transformers>=4.36.0",
    
    # LLM providers
    "google-genai>=1.0.0",
    "httpx>=0.26.0",
    
    # Agent framework
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
